Revert "Add ReasoNet model as example"
This reverts commit 09daade4f080d930af51f78508769bd589d25bc9.

+++ Source/CNTKv2LibraryDll/API/CNTKLibrary.h
@@ -364,7 +364,7 @@ namespace CNTK
                 if (i != 0)
                     wStrStream << L" x ";
 
-                wStrStream << ((int)m_shapeDims[i]);
+                wStrStream << m_shapeDims[i];
             }
 
             wStrStream << L"]";
